+++ docs/CHANGELOG.md
@@ -15,6 +15,10 @@
 * New package `cosineschedule`, refactored from `optimizers` package.
   * Added handling negative values for the hyperparameter `cosine_schedule_steps`: they set the period of the cosine schedule
     as fractions of the total number of steps being trained.
+* Package `train`:
+  * Extensions to `Dataset` interface through additional interfaces.
+  * Added optional `IsOnwershipTransfer() bool` that allows a Dataset to specify it should maintain ownership of the 
+    yielded tensors.
 * Updated `gopjrt` v0.4.4 with the static XlaBuilder library, and experimental support for Apple/Metal.
 
 ## v0.13.0 - 2024/10/07